Add files via upload
This contains a table .Each column describes how many games have been won by each team (home/away) and based on that we are predicting which team has won or lost. We have a set of train datas in which we fit the model ,then, we are comparing the results with test data. And we are finding the accuracy score