Add files via upload
Today , many files in our gmail doesn't get recognized as SPAM even though we might not need it.We have to manually send it to a SPAM folder. Wouldn't it been nice if we can write a code and it will perform the prediction of chunk of emails automatically whether it is a  SPAM folder!!!! Hence i made a try to do this:This ipynb is used to provide an algorithm which helps to connect to your email and fetch all the mail subjects  that you provide such as inbox, spam and other repositories . After doing this it predicts whether the file is actually SPAM or not.

Left: Now i have to pro-grammatically migrate all the emails that are identified as SPAM. I have not yet got the desired prediction and i am in the hunt of it.